cancel command

+++ src/main/java/logic/commands/maincommands/SaveCommand.java
@@ -6,11 +6,11 @@
 package logic.commands.maincommands;
 
 import logic.configuration.ConfigurationManager;
+import logic.database.AttachmentDAO;
 import logic.database.EmployeeDAO;
-import logic.entity.Address;
-import logic.entity.Attachment;
-import logic.entity.Employee;
-import logic.entity.Photo;
+import logic.database.PhoneDAO;
+import logic.database.PhotoDAO;
+import logic.entity.*;
 import logic.processcommand.ActionCommand;
 
 import javax.servlet.http.HttpServletRequest;
@@ -21,6 +21,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Iterator;
 
 import static logic.configuration.LogConfiguration.LOGGER;
 
@@ -40,10 +41,12 @@ public class SaveCommand implements ActionCommand {
     }
 
 
-    public synchronized boolean saveContact(HttpServletRequest request) {
+    public boolean saveContact(HttpServletRequest request) {
         EmployeeDAO contactDAO = new EmployeeDAO();
         Employee employee = getEmployeeFromSession(request);
         updateEmployee(request, employee);
+
+        savePhones(employee.getPhoneList(),employee.getId());
         saveAttchment(employee);
         if (!employee.getPhoto().isExistInDB() && employee.getPhoto().getPhotoName() != null) {
             savePhoto(employee.getPhoto());
@@ -60,13 +63,45 @@ public class SaveCommand implements ActionCommand {
         return true;
     }
 
+    public boolean savePhones(ArrayList<ContactPhone> phones,final int EMPLOYEEID){
+        Iterator<ContactPhone> phoneIterator = phones.iterator();
+        PhoneDAO phoneDAO = new PhoneDAO();
+        while (phoneIterator.hasNext()){
+            ContactPhone phone = phoneIterator.next();
+            if(!(phone.isSaved() || phone.isDeleted() || phone.isUpdated())){
+                phoneDAO.addPhone(phone,EMPLOYEEID);
+            }
+            if(!phone.isSaved() && phone.isUpdated()){
+                phoneDAO.updatePhone(phone);
+            }
+            if(phone.isDeleted() && !phone.isUpdated()){
+                phoneDAO.deletePhone(phone.getId());
+            }
+        }
+        return true;
+    }
+
     public boolean saveAttchment(Employee employee) {
+        AttachmentDAO attachmentDAO = new AttachmentDAO();
         ArrayList<Attachment> attachments = employee.getAttachmentList();
         String filePath = ConfigurationManager.getPathProperty("path.saveFile") + employee.getId() + "/";
         for (Attachment attachment : attachments) {
-            if (attachment.isDeleted()) {
+            if (!(attachment.isSaved() || attachment.isDeleted() || attachment.isUpdated())) {
+                attachmentDAO.addAttachment(attachment);
                 String resultFileName = filePath +
-                        attachment.getFileName();
+                        attachment.getId();
+                System.out.println("RESULTFileName: " + resultFileName);
+                try {
+                    Path path = Paths.get(resultFileName);
+                    Files.write(path, attachment.getAttachment());
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+            if (attachment.isDeleted() && !attachment.isUpdated()) {
+                attachmentDAO.deleteAttachment(attachment.getId());
+                String resultFileName = filePath +
+                        attachment.getId();
                 Path path = Paths.get(resultFileName);
                 try {
                     Files.delete(path);
@@ -74,23 +109,26 @@ public class SaveCommand implements ActionCommand {
                     LOGGER.error("can't delete file from server " + e);
                 }
             }
-            if (!attachment.isSaved()) {
+            if(!attachment.isSaved() && attachment.isUpdated()){
+                attachmentDAO.updateAttachment(attachment);
                 String resultFileName = filePath +
-                        attachment.getFileName();
+                        attachment.getId();
                 System.out.println("RESULTFileName: " + resultFileName);
                 try {
                     Path path = Paths.get(resultFileName);
-                    System.out.println("Path: " + path.toString());
                     Files.write(path, attachment.getAttachment());
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
             }
+
         }
         return true;
     }
 
     public boolean savePhoto(Photo photo) {
+        PhotoDAO photoDAO = new PhotoDAO();
+        photoDAO.updatePhoto(photo);
         String resultFileName = ConfigurationManager.getPathProperty("path.saveFile") + photo.getEmployeeID() + "/photo/";
         File uploadDir = new File(resultFileName);
         if (!uploadDir.exists()) {
@@ -99,7 +137,6 @@ public class SaveCommand implements ActionCommand {
         resultFileName += photo.getPhotoName();
         try {
             Path path = Paths.get(resultFileName);
-            System.out.println("Path: " + path.toString());
             Files.write(path, photo.getBytes());
         } catch (IOException e) {
             LOGGER.error("can't write photo on disk: " + e);
@@ -109,7 +146,11 @@ public class SaveCommand implements ActionCommand {
         return true;
     }
 
+
+
     public boolean deletePhotoFromDisk(Photo photo) {
+        PhotoDAO photoDAO = new PhotoDAO();
+        photoDAO.updatePhoto(photo);
         String resultFileName = ConfigurationManager.getPathProperty("path.saveFile") +
                 photo.getEmployeeID() + "/photo/" + photo.getPhotoName();
         Path path = Paths.get(resultFileName);
@@ -155,11 +196,6 @@ public class SaveCommand implements ActionCommand {
             address.setIndex(request.getParameter("house").isEmpty() ?
                     0 : Integer.valueOf((request.getParameter("index"))));
         return address;
-    }
-
-    public boolean isNewEmployee(HttpServletRequest request) {
-        Employee employee = getEmployeeFromSession(request);
-        return employee == null;
     }
 
     public Employee getEmployeeFromSession(HttpServletRequest request) {
